w3c-reports/12092014/gtk/runned-tests/html/syntax/parsing/html5lib_adoption01-expected.txt

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

html5lib Parser Test


FAIL html5lib_adoption01.html dab5eca760a630bc57719d678d789dd1ca74f492 null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_adoption01.html a3a46907dc73b7be1e1171f797a9f696b7fb185b null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_adoption01.html b720cd1d95283d6288e7ca17142540b10ef8f847 null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_adoption01.html 83b2bd6f0849b2c7283448f15095806a17c7f0c0 null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_adoption01.html 1d56fa2d9d756f9053638a3db25fcd84e57e41ec null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_adoption01.html c129fce97933067558d7833632ad7ef2d149616e null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_adoption01.html 39a3ec04f54cda2bf1e06c54caf0b302da4fd252 null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_adoption01.html d0bdfe6be48309b2d01b497667b350a8e6ec8ffb null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_adoption01.html 63401082c6afd4b3c3201e348a1a61722f167b21 null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_adoption01.html 883cfd89f63da854dffbfbc938da1f31887a55e5 null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_adoption01.html 4dc0665051dac0d72c71aba5c95c5b86437dea2c null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_adoption01.html 4f376c0a798e71a91065e215dc0175d3107d5208 null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_adoption01.html 52b62611a847a3f5fc3dc607a6b0174f1697247c null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_adoption01.html 88d0d3403d2a7b4058fcfb4e62835acb1c207e0f null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_adoption01.html d7338f457789f65d47b240e203b9e40a3925f2ca null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_adoption01.html bf8088acd8bd48d8487bae49a613c248c488c041 null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
FAIL html5lib_adoption01.html 39edaa2e298c60ad4a1b0b9fdb4481d4aea98f36 null is not an object (evaluating 'element.normalize')(stack: test_serializer@http://localhost:8800/html/syntax/parsing/test.js:31:10
http://localhost:8800/html/syntax/parsing/test.js:162:52
step@http://localhost:8800/resources/testharness.js:798:30
onload@http://localhost:8800/html/syntax/parsing/test.js:160:13)
